delete gateway

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -43,9 +43,14 @@ const Mutations = {
     async deleteItem (parent, args,ctx, info){
         const where = {id: args.id}
         // find item
-        const item = await ctx.db.query.item({where}, `{id title}`);
+        const item = await ctx.db.query.item({where}, `{id title user {id}}`);
         // check if they have permission for that item
-        // TODO
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermission = ctx.request.user.permissions.some(permission => ["ADMIN", "DELETEITEM"].includes(permission));
+    
+        if(!ownsItem && !hasPermission){
+          throw new Error("You don't have permission to do that");
+        }
         return ctx.db.mutation.deleteItem({where}, info);
         // delete item
     },
